requirements.txt
# Runtime GUI
PySide6>=6.7,<7
python-dotenv>=1.0,<2
openpyxl>=3.1,<4

# Tests
pytest>=8.0,<9
